Node.js sample file for mobile push notifications
This a NodeJS model to send a push notification to an iOS device or Android device. The Node.js applications main purpose is for P2P payment processing, instant messaging, and profile viewing. 
Requires an object of the push payload, and the recipients device token. Is used to send a notification to the payer, and the payee